requirements.txt
PyYAML>=6.0
rich>=13.0.0
click>=8.0.0
inquirerpy>=0.3.0
pytest>=7.0.0